Manifests.Update() on fresh Manifests copy
The operator is performing multiple updates on the same manifests instance - IOW the Update() function processed a dirty copy of Manifests object if more than a single Update() call being executed.
This led to multiple bugs which I initially thought their source is in the deployer:
1. https://github.com/k8stopologyawareschedwg/deployer/pull/55
2. https://github.com/k8stopologyawareschedwg/deployer/pull/56

In order to solve this issue we change the Update() logic to always work on a fresh copy of Manifests object, hence we won't carry former changes during the Update() call.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>