Consume RTE e2e tests
The RTE e2e tests are using the k8s test framework (for compatibility with NFD) but in the the numaresources-operator tests it is more conveniente fo us to use the controller-runtime.
But k8s test framework's clientset and controller-runtime clientset can't live together.

In order to solve this issue we're compiling the RTE specific tests as a separate binary and running the binary as part of the numaresources-operator tests.
By having two separate binaries we're avoiding the conflicts between the different clientsets.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>