rte: create configmap from kubeletconfig
revamp how we handle the configuration of the RTE.
We need to supply to the RTE a configuration file to tell it:
1. details about the TM configuration, which it cannot figure out
   otherwise (policy, scope...)
2. parameters needed by the `sysinfo` package, which it cannot
   determine from the system because they depend on the kubelet
   configuration.

The second set is going to be phased out when OCP exposed the
latest changes in the podresources API, but the first set will
always be needed. So we will need some form of configuration management
anyway.

So the operator will translate the information from the Kubeletconfig
object, if available, to the RTE config.yaml and make it available as
configmap.

Most important design decisions:
a. always translate as much configuration as possible, since it's cheap
   once we have access to the relevant kubeconfig. This for example
   include always reporting `reservedCPUs`, even when the RTE will no
   longer need this information (podresources' GetAllocatable
   available). RTE will just ignore the fields it doesn't need anymore
b. build the configuration only on OpenShift; on Kubernetes, we can
   just mount the kubelet config on the RTE pod and let it figure out.
c. on the other hand, always mount the configMap, marking it `optional`,
   so we will not need platform-dependent changes on the DaemonSet.

Summary: always try to mount the configMap from the RTE daemonset,
ignore errors if not available, create it only on OCP translating as
much as possible (greedily).

In order to process the kubeletconfig, we introduced a new controller.
The new controller will watch the kubeletconfigs and reconcile the
relevant configmaps.
We add a new controller to not overload the existing controller,
and because, per above design decisions, we can nicely split the
responsabilities and make the two reconciliation loops
completely independent: the numaresources main controller will
deploy the bigger part of the stack; the kubeletconfig controller
will provide configuration, if needed and possible.
(xref: "One controller per custom resource definition" from
https://sdk.operatorframework.io/docs/best-practices/best-practices/)

To handle the case on which the configmap is created after the
daemonset, we make the configmap mount mandatory, and we let the core
reconciliation loops of kubernetes eventually fix any issue.

Finally, a later patch will make sure the RTE itself can pick
up ConfigMaps updates automatically.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>
Signed-off-by: Francesco Romani <fromani@redhat.com>