rte: use operand image as rte image
Revamp the code handling user-supplied RTE image, to streamline
the flow and make room for future enhancement needed to fully
fix the operator-as-operand flow.

Expose and adjust the environment variables we use to introspect the image.

Important fix: We need to passthrough the ImagePullPolicy
in the operator-as-operand flow; this is correct in general,
the image needs to be handled consistently, and unbreaks our
CI flows.

Signed-off-by: Francesco Romani <fromani@redhat.com>