rte: bump UID/GID when reusing operator image
The RTE need to access entries in the filesystem which
are owned by root - surely the podresources socket,
on k8s also the kubelet config.

However the operator should run as non-root,
and its configuration reflects so.

When we use the operator image as operand (RTE), this settings
"leaks" (positively for a change!) into RTE.

The access to the files (podresources socket) needed by RTE
is guarded by both SELinux and unix permission control.
We already fixed the former when needed, but to pass the latter,
we need to run RTE with uid and gid 0 (as root).

Note we do NOT want full root privileges: we rely on general
kube settings to drop all the other privileges; should unneeded
privilege not have been dropped, it will be fixed on separate PRs.

Signed-off-by: Francesco Romani <fromani@redhat.com>