controllers: nro: simplify isMachineConfigPoolsUpdated
simplify the helper function to take the minimal arguments;
the goal os this refactoring is to allow this function
to be used later outside the controller.

Introduce an explicit helper to synchronize the relevant
MCP statuses into the NRO object status.

Overall, code movement and remix, no intended changes in behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>