RTE: add option to exit if the config file changes
RTE needs to react to config file changes.
There are two ways: either a massive refactor of the internal
flow to detect and consume the config changes, or add an
option, disabled by default, to just exit when the a config
change is detected (with a log emitted to signal that), to
let the supervisor restart the process with the updated config.

Since we expect few and far between config changes, we start
with the simplest option.

Signed-off-by: Francesco Romani <fromani@redhat.com>