nro controller: use predicates for all resources
Our apply logic compares the whole object and also update the whole object
including metadata and status that can be a reason for additional reconcilation
loops. Instead of improving the comparison and update logic that will make
the code more complex, we can just filter events for objects that did not
have any change under spec or labels.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>