e2e: move install into a separate test suite
We need to control the order of the NUMAResourcesScheduler tests:
1. Install
2. Other tests
3. Uninstall

For that we need each one of the above to be placed in a separate suite and call the suites in the desired order.
This commit is contains the first step which is refactoring the existing install test and place it in a separate suite.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>