images: remove defaults
Previously, we had a safe fallback for the RTE image.
But we don't support this flow anymore; the recommended
and supported flow is to use the same operator image as RTE image,
and to optionally allow (in rare and specific circumstances)
the cluster admin to override the RTE image.

Additionally, we must detect now the case on which we have neither
detected the operator image using introspection, nor we have
a user-supplied image.

In this case, most likely caused by a manifest/configuration error,
we must mark the CR state as degraded.

Signed-off-by: Francesco Romani <fromani@redhat.com>