Makefile: use the most updated tag as default
When a user is running ```make deploy``` without specifiying any specific option we should deploy the controller with the most updated version.
The most updated version that is getting build (automatically by the CI) is with the 4.10-snapshot tag, hence that should be the default.

Not running with the most updated version out of the box, might lead to gaps between what specified in the manifests (flags, args, etc.) and what the binary in the image actually contains

Signed-off-by: Talor Itzhak <titzhak@redhat.com>