flagcodec: explictely split Command and Args
Instead of assuming that argv[0] is the command, change the
API to require command and args to be explicitely set, for
better clarity.
Extend the package to return the processed args, or the full
argument vector (command + args, only the latter actually processed
and the former passed through verbatim) for the sake of the
completeness.

Signed-off-by: Francesco Romani <fromani@redhat.com>