e2e: bootstrap serial lane
The "serial" lane will have e2e tests which check
the whole scheduler + RTE solution, thus aiming to validate
the real e2e, cluster level solution and not just the
individual components.

For this reason, each test need to change the cluster state,
even though is responsability of each test to clean up everything.

Hence, each test is potentially disruptive and must run
in complete isolation with respect any other cluster activity,
not just any other test.

Mimicing(ish) what k8s upstream does, we collect these tests
in a "sefial" suite, to be called in a strictly serial and exclusive
(as in mutual exclusion) way.
The latter has to be addressed using proper cluster and test runner
configuration, hence it is out of scope for this PR.

Signed-off-by: Francesco Romani <fromani@redhat.com>