e2e: introduce NO_TEARDOWN option
In order to explain the motivation of this PR let's define the following terms:
Clean state: The cluster state, before any MC changes and/or TAS components were applied.
Dirty state: The cluster state, after any/all MC changes and/or TAS components were applied.

In the ususal test cases, we would always want the cluster to be at clean state, after the test run.
But there are special cases, like local tests and GH action tests, that we would like to keep the cluster in dirty state so we could perform various investigation, exporting logs, events, etc.

For that, we'll introduce the NO_TEARDOWN option, that will allow us to keep the cluster in dirty state after test run.
This is totaly optional and keep the current behavior intact.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>