gh actions: opt in the kind e2e flow
We always considered the kind e2e flow a extra sanity check
and a tier-2 platform because this operator has -and will have-
a number of unavoidable OCP dependencies.

Since the kind e2e lane is flaking recently and no good
resolution is in sight, we make the kind e2e-lane opt in.

We totally want to keep deploying on kind, and we still
recommend to test also on kind.

gh docs references:
https://docs.github.com/en/actions/learn-github-actions/contexts#github-context
https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#push

inspired by:
https://github.com/veggiemonk/skip-commit/issues/5

Signed-off-by: Francesco Romani <fromani@redhat.com>