e2e: utils: add function to create test pods
even though running bare pods is a kube antipattern,
submitting pods is useful for fine-grained testing.
We start with the simplest function to create a bare-bones
pause container pod.

We will evaluate if richer pod builder infra is needed,
but the return of investment is not clear yet, so
it seems beneficial to start lean generalizing the common
parts and specializing the pod on a test-by-test basis.

Signed-off-by: Francesco Romani <fromani@redhat.com>