pkg: create new 'objectnames' package
The `objectstate/rte` package was holding code to create names
for objects managed by the operator.
We initially put this code here out of practicality, but
this code doesn't belong here; actually, it doesn't belong in
any of the existing packages.

So we create a new package whose responsability is to create
the names. This helps also in de-entangling our internal package
dependency tree.

Only trivial code movement and renames.

Signed-off-by: Francesco Romani <fromani@redhat.com>