flagcodec: preserve the order of args
Up until now, `flagcodec` was reordering the args
before rendering them.
This is not always correct, because it's relatively common
for command line argument to have some ordering requirements.

A key example is the subcommands: some options for a command
may make sense for the global context, while some other options
are subcommand-specific.

flagcodec doesn't aim (yet) to cover all the cases, but this is
simple to fix: instead of force a sorting, just record the order
on which the flags are parsed, and reuse it when rendering.
The end result is still a valid command line, the implementation
is not more complex and the behaviour is more correct.

Signed-off-by: Francesco Romani <fromani@redhat.com>