e2e: add numacell dp for testing purposes
Add device plugin to expose the numa cells of a machine as resources,
through the device plugin API of the kubeket.
Includes the related manifests + go packages to interact with them.

This patch enables the e2e tests to pin specific numa cells on
specific machines, greatly sinplifying the test fixture setup.

This approach breaks badly the design of kubernetes and its spirit,
so it should NEVER EVER be used outside e2e tests and on any cluster
outside CI and (early) development stage.

Signed-off-by: Francesco Romani <fromani@redhat.com>