e2e: fix potential race between controller and assertion check
In the current behavior we're changing the ```LogLevel``` field under the CR and then check the change results under the container.
There might be times when the reconciliation loop haven't occur yet (thus the pod didn't get updated) and we're already checking for the pod changes under the test.

With the new behavior using Eventually we'll have until timeout to check if the test condition were met.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>