e2e: install: wait for objects to be deleted
the install suite was flaking with errors similar to
```
Message: "object is being deleted: kubeletconfigs.machineconfiguration.openshift.io \"kc-test\" already exists",
```

This is because the test code was not waiting for the object
to be deleted from the cluster, but it was rather going along
if the deletion *request* was succesfully accepted by the system.

It's more correct to wait for objects to be deleted before to
keep going with the tests, so we do it in this patch, adding
the necessary utilities to enable this flow.

Signed-off-by: Francesco Romani <fromani@redhat.com>