e2e: serial: test multi-container placement
Add e2e test to verify the behaviour of the NUMA aware scheduler
by creating a testcase which is expected to foul the default,
NUMA-unaware scheduler.

We expect to run on a cluster with two or more machines each
of which has two NUMA zones. We don't actually support
more than two zones because the resource allocation numbers
are hardcoded, not computed on the fly. Given how rare machines
with more than 2 NUMA zones, that should not be a problem anytime soon.

We use a guaranteed pod with two containers; we set the required
resources such as the total amount of requested resources is available
on both thw two workers but one of them only has all resources on one
numa per container.

IOW:
- 2+ nodes have all the required resources availaible at *node* level
- exactly 1 node have all the required resources available at *NUMA
  zone* level.

Hence we expect here the NUMA-aware scheduler to pick the only right
spot every time, at the first try.

Signed-off-by: Francesco Romani <fromani@redhat.com>