Conditions avoid update due to wrong timestamp
We create a new `default` set of conditions to compare with the current
ones. These conditions are created with `now()`	as value for
`LastTransitionTime` field.
This way comparison always return there are differences between
conditions and therefore they are always updated even if it is not
needed.

Update `LastTransitionTime` to have the same value as the current
conditions so it would not modify the comparison.

Note: everything in condition handling is based on the assumption that
the conditions order in the conditions array is always the same.