e2e: serial: generalize test multi-container placement
Generalize the existing e2e test which verifies verify the behaviour of the NUMA aware scheduler
by creating a testcase which is expected to foul the default, NUMA-unaware scheduler:

- 2+ nodes have all the required resources availaible at *node* level
- exactly 1 node have all the required resources available at *NUMA
  zone* level.

Hence we expect here the NUMA-aware scheduler to pick the only right
spot every time, at the first try.

We intentionally just check if the TM policy matches expectations, avoiding implicit
restart: is up to the test runner to setup the config accordingly, orchestrating reboots,
and perhaps running the same suite twice.

Signed-off-by: Francesco Romani <fromani@redhat.com>