e2e: serial: use generateName for padding pods
Padding pods are actually anonymous: we use node+zone in
their name only to make easier to easily find which one
pertains to which node, but that can also backfire because
the hostname can make the pod name exceed the limit.

So we switch to generateName and we add node/zone name as label,
which should be good enough to crosscorrelate the padding pods.

Signed-off-by: Francesco Romani <fromani@redhat.com>