numaresourcesoperator: configmap hash annotation
We're adding an annotation with the ConfigMap's hash which is used by RTE, under the DaemonSet .spec.template filed.
This way, for every change/update done to the ConfigMap, a new hash will get calculated which in turn causes the DaemonSet to restart the RTE pod and read the updated ConfigMap.
This change is important in order to make sure that we're not running RTE pod with a stale ConfigMap configuration.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>