numaresourcesscheduler: configmap hash annotation
We're adding an annotation with the ConfigMap's hash which is used by the scheduler, under the deployment .spec.template filed.
This way, for every change/update done to the ConfigMap, a new hash will get calculated which in turn causes the deployment to restart the scheduler pod and read the updated ConfigMap.
This change is important in order to make sure that we're not running the scheduler pod with a stale ConfigMap configuration.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>