e2e: serial: start the non-regression test group
start splitting the non-regression tests into their own go source.
Non-regression tests are tests which guarantee the numa-aware scheduler
behaves like the original scheduler in cases on which numa-awareness
is not needed or relevant.

As side benefit of this move, we de-entangle a test by a unwarranted
requirement inherited from the previous layout, enabling it to run
against any TM scope.

Signed-off-by: Francesco Romani <fromani@redhat.com>