e2e:serial: modify configurable values test
The test modifies every configurable value under the NUMAResourcesOperator and NUMAResourcesScheduler CR's and checks that the controller applied the configuration as expected.
The test requires MCP changes, hence nodes reboots and therefore it's label with a special [reboot_required] label to denote that.
We also add the [slow] label to denote this test might takes more time than usual.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>