e2e: serial: start the overhead test group
split the "overhead" tests into their own go source.
"overhead" tests are tests which want to check how the system deals with
the pod/container overhead, either managed by the platform or implicit.

As side benefit of this move, we de-entangle a test by a unwarranted
requirement inherited from the previous layout, enabling it to run
against any TM scope.

Signed-off-by: Francesco Romani <fromani@redhat.com>