sched: yaml: reorder filter chain
Move the NodeResourceTopologyMatch as early as possible in the filter
chain.
The scheduler framework runs for each pod all the configured filter
sequentially, bailing out at the first non-success (pods _may_ be
processed in parallel between each other).
Previously, the NodeResourceTopologyMatch plugin was enabled and
(implicitely) put at the end of the chain. We can observe this looking
at the config the scheduler dumps in the log (with -v=6 at least)
at startup.

When things work correctly, it doesn't make any difference where
the plugin is in the chain, because all of the filters must pass.

When we need to troubleshoot things however, it's useful to have
our plugin - which is also the less battle tested atm - as early as
possible because we can both have better control (greater chances
the plugin run at all instead of being skipped because some other
filter unexpectedly failed) and we can add custom logs to peek into
the scheduler behaviour.

To implement this change, we need to explicitly disable all the filters
and re-enable them in the order we desire
(see:
https://kubernetes.io/docs/reference/scheduling/config/#extension-points).

We take this chance to DISABLE some cloud-provider-related filters,
which are not relevant for the use case this operator wants to cover.

Signed-off-by: Francesco Romani <fromani@redhat.com>