e2e: serial: take into account base load
the interaction between the noderesourcetopology and the noderesource
filter plugins highlighted quite few complications.

by design, NRT objects do NOT account for non-esclusive resources;
this mean burstable pods do take node resources, but these are not
reflected into NRT available fields. This is unfortunate, but due
to the fact that non-esclusive resources are not pinned to a NUMA
zone, hence are difficult to track and account.

This means the NRT and the noderesource plugin have a different
perspective about the available resources.
Note the kubelet has the same picture as the noderesource plugin
(bar network delays) hence
- noderesource can cause a node to be filtered out "unexpectedly"
- just disabling noderesource (debatable) can lead to pod rejections
  for OutOfCpu or OutOfMemory errors, postponing but not solving the
  issue.

All the above means that tests which want to saturate a node
must take the burstable pod into account and not assume eviction
will solve all the problems, which is what we do in this PR by adding
(estimate of) node base load and using these values when doing the
node padding.

Additionally, to improve the troubleshooting experience, we
can and should improve the scheduling configuration, but that's
being handled in different PRs (e.g.
https://github.com/openshift-kni/numaresources-operator/pull/220 )

Signed-off-by: Francesco Romani <fromani@redhat.com>