e2e: serial: introduce the `config` subpkg
Move the configuration steps we used to do in the
{Before,After}Suite{,Helper} in their own subpackage.

This way, we can keep the current suite behaving like before
(bar the necessary mechanical API and import changes) but we
enable better integration with external projects (= cnf-tests)

It goes like this:
cnf-tests want to run the configuration steps in a different
suite (configsuite) separated by the test suite proper (e2esuite).

In this model, we will add all the configuration steps in the
config suite; in the test suite, we will integrate the tests
proper and the minimal configuration support (our own fixture).

Signed-off-by: Francesco Romani <fromani@redhat.com>