e2e: runner: add --report-file option, rename to --report-dir
The `--report-dir` (was `--report`) is useful when we run with
`--setup` and/or `--teardown` and we want to collect all the reports
in the same directory.

However a common use case is to just run the tests, and for that
`--report-file` to directly set the report file location may be handy.

Signed-off-by: Francesco Romani <fromani@redhat.com>