project: add the internal/ hierarchy
Our e2e tests are complex beasts which require a good amount of support
code. Some of this support code could in turn use some unit testing
to reduce the testing tuning churn, but it's awkward to have tests
inside tests.
On the other hand, most of this code, albeit useful, feel awkward to
be placed under pkg/ because this is not code we really want to export
and support.

To move forward, we introduce the internal/ hierarchy which sits in
a pretty nice middle ground. We can have in here utility packages which
should be almost pkg/ quality, but which we don't need to support -
the tooling will prevent external users.

See: https://dave.cheney.net/2019/10/06/use-internal-packages-to-reduce-your-public-api-surface

We may want to do more code movement from/to internal/ in the future,
and we may want to revisit our pkg/ tree.

Signed-off-by: Francesco Romani <fromani@redhat.com>