e2e: fixture: add cooldown on teardown
The e2efixture deletes the namespace it manages and
wait for it to be gone, to try hard to leave the cluster
as clean as it found, and to make tests independent from
each other.

Trouble is: the accoubting done by RTE/NFD usually take
nonzero time to react to container deletion and to reclaim
resources, so tests can race with NRT and get false
positives/negatives.

The proper fix would be to record all the NRT values and
wait at the very end of the fixture for the values to be
restored to the pristine values.

This is hard to do, and takes time to get it done.
A (crude) approximation is to just wait "some time".
This reduces the chances of flakes and it's still a
step forward.

Signed-off-by: Francesco Romani <fromani@redhat.com>