e2e:serial: unschedulable numa level test
This test creates a deployment with multiple replicas with a pod that is not fit into a node at the numa-level, meanning that there are enough resources in all numa nodes together, but not enough in each one of them, so all pods` replica scheduling must be failed.

In the next step we're reducing the amount of resources requests by the pod and expect at least some of the replicas to be running.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>