e2e: serial: config: add Ready() helper
Now that potentially we can run without assurance that
the setup was NOT done properly (this should never happen,
but at the end of the day we have less guarantees), we can
use some extra protection.

So we add a prerequisite on the suites which actually depend
on the fixture data being initialized, to fail with a clear
message if the preconditions aren't met.

Signed-off-by: Francesco Romani <fromani@redhat.com>