e2e: serial: utils: helper to wait for padding pods
Factor out common code in tests, especially e2e tests, is actually
a complicated business.

Test code has a different set of tradeoffs and goal wrt the application
code, so the usual rules and best practices don't trivially apply.
A good reference material is https://about.sourcegraph.com/go/advanced-testing-in-go
- as the very introduction of the page states, this is more
a conversation starter and food for thought than a hard reference.
Still, we need some references.

So, why introduce a new helper? does it conflict with (admittedly
controversial) guidelines in
https://about.sourcegraph.com/go/advanced-testing-in-go/#Repeat-yourself ?

Thing is, observing ex-poste the patterns in our tests, setting up the
padding pods to recreate a specific allocation factor on nodes, is
indeed "very reused logic that doesn’t fail often (example: changing a directory)
or fails all at once (creating a test server)". This step should indeed
never fail, and if it fails it catastrophic.

So it makes sense to consolidate all the patterns, including post-mortem
event log, in a common helper.
Additionally, consolidating this specific patterns allows us to fix
the annoying excessive output issue (due to the fact we Expect() a
collection of big pod objects to be empty) which hampers the
debuggability. We can now consolidate in a terser yet equivalent from
debug purposes expectation.

Signed-off-by: Francesco Romani <fromani@redhat.com>