e2e: serial: add option to dump events
Add a tunable, using a environment variable, to dump
to stdout the events pertaining to a failed pod.
Previously, we just logged this information.
Logging is fine (and it's the most correct option)
but logs are (again correctly) sent on stderr, which is sometimes
not readily accessible on CI systems.

By adding the option to dumping the events, we make troubleshooting
easier on these CI systems.

Signed-off-by: Francesco Romani <fromani@redhat.com>