e2e: serial: consistent naming for all the env vars
Let's use consistent names for all the env vars that
allow to tune the behaviour of the serial suite.
Still honor the old names for backward compatibility and
minimal disruption

This is still a WIP while we identify good and consistent
names across the board, this is why the docs aren't updated yet.
Even when we will recommend the new names, we will still keep
considering the old names for backward compatibility for some time,
with extra warnings added.

Signed-off-by: Francesco Romani <fromani@redhat.com>