e2e: serial: taint: more careful untaint
We had a string of issues related to taint added in tests lingering
around and polluting the cluster state.
To make the test more robust, we store the taint applied and the
node we tainted, and we make sure to clean them first.
This is not more fragile than before because it has the same
failure model (state lost) if the test is interrupted or crashes
too early.

After we clean up after ourselves, we add a extra pass iterating
over the worker nodes (logging loudly along the way) to make sure
to remove any taint that could have leaked.

The extra safety comes from this defensive extra step.

Signed-off-by: Francesco Romani <fromani@redhat.com>