e2e: serial: taints: wait for taints to be gone
We had a string of issues related to taint added in tests lingering
around and polluting the cluster state.
So, at the very end of the test adding the taints, we add an
explicit wait loop to wait for any taint still possibly lingering
around to be gone before to claim the test is finished.

This should at least make leaks more evident.

Signed-off-by: Francesco Romani <fromani@redhat.com>