e2e: serial: test alignment of each single resource
Add tests to check the scheduler behaviour when a single
resource is the deciding factor. These tests are
intentionally a close variant of the single-pod single-container
tests which we have already and which were incidentally checking
this flow.

In this test group we want to have explicit and more extended
coverage about the scheduler decision with a single resource
as deciding factor.

Signed-off-by: Francesco Romani <fromani@redhat.com>