Numaresource Operator Serial Test Improvements
This patch is make the test suite more robust.
Rather than using hardcoded scheduler name, we obtain it from
the numaresourcesscheduler CR.

In addition to that, we perform a check to make sure that in case
pod event checks return no events, we print a warning as that could
indicate misconfiguration in the cluster.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>