e2e: wrap the api client operations in Eventually()
Applying the API client operations like Create/Delete/Update may take time and if not waited enough the outputted error could end up not nil, which in turn fails the test.

Wrapping such operations in Eventually or loop will guarantee that moving to the next test steps is done only after waiting a while (2 mins), hence stabilizing the output of the operation.

Signed-off-by: shereenH <shajmakh@redhat.com>