introduce the scheduler profile `*-sorted`
Add and automatically maintain a `-sorted` variant
of the scheduler profile.
"Maintain" here means that we make sure the `-sorted` variant
always follows the name the user sets. If the user sets
the scheduler name `custom-sched`, the controller will make
sure the variant will be named `custom-sched-sorted`.
If the name is later changed `awesome-sched`, the variant
will be then renamed `awesome-sched-sorted` automatically, and so on.

The `-sorted` variant wants to run the very same set of filter as
the custom variant, but with a different order of filters.
The main benefit is the `-sorted` variant runs the NRT filter *BEFORE*
the `noderesourcefit` filter.

The `-sorted` variant is not explicitely advertised but it's
supported, and it is expected to be exercised by tests (e2e most
notably).

Signed-off-by: Francesco Romani <fromani@redhat.com>