build: introduce buildhelper replacing git-semver
Replace git-semver, which we were using in the build step,
with another tailor-made tool superset of it.
What we gain is: we can now report the git commit from which
the build is done, and we can inject the values from outside,
bypassing the autodetection, to be friendlier with isolated build
systems, which would otherwise fail the autodetection.

Signed-off-by: Francesco Romani <fromani@redhat.com>