rte: add helper to find containers
Previously, we were blindly trusting the order of
RTE containers to apply our updates. Now we do an explicit
search based on the container name.

This is not the definitive fix because the names (and some searching
facilities in general) should be provided by the deployer package,
but it is still a small step in the right direction of code more robust.

Signed-off-by: Francesco Romani <fromani@redhat.com>