e2e:serial:tmpol: generelize checking for consumed resources
After the test pod is landing, we're checking the NRT objects to see if resources were consumed accordingly.

In ```SingleNUMANodeContainerLevel``` the resources for the pod MIGHT be consumed from multiple zones, so the current ```CheckZoneConsumedResourcesAtLeast``` function is not good enough.
The ```CheckNodeConsumedResourcesAtLeast``` should handle the check correctly.

We provide a function pointer to the relevant function as a new parameter for the test.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>