e2e:serial:tmpolicy: add scope=container tests
Add tier1 tests which verify the core functionality of the scheduler when nodes are configure with scope=container.
The tests types` are taken from https://github.com/kubernetes-sigs/scheduler-plugins/blob/master/pkg/noderesourcetopology/TESTS.md

We took the liberty to add the most critical and diverse types of tests
among the tier1 ones.
We can add more later if we find it necessary.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>