pkg: mcp: introduce `NodeGroupTree`
Add the concept of `NodeGroupTree` which represents the binding between
a NodeGroup and all the MachineConfigPools it refresh. The relationship
can be 1:N - but worth remembering we cannot have MCPs belonging to more
than 1 NodeGroup.

So, add a new type and the related functions to get all the relevant
NodeGroupTrees from a slice of NodeGroups.

NodeGroupTrees will turn out handy in a later change, where we will
do transformation on objects derived from MCPs (daemonsets) depending
on settings of their parent NodeGroup.

Signed-off-by: Francesco Romani <fromani@redhat.com>